shopping-backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
yarl==1.8.2
